Add Multi Biome Support (#36)
* Add the new auto blend layer

In addition to that, we now fixed an issue with edge mask in the auto blend. Earlier it mixed UVs using a lerp. Usually, that causes issue in between values.

* Use a Smart Mask as the snow layer.

* Allow smart masks to override nodes.

* Added the new master material.

* Add the snow layer.

* Make it ready to try out

Now the example map is created using the Multi-Biome support. So, now we can play with this.

* Get changes from the main branch.

Also, now our material looks like the same as the old version.

* Add all the defaults.

* Use Multi-Biome with the Mobile Map

* Add Support to Mac.

* Allow to grow grass on auto layers.

* Add grass support for multi-biome

* Add multi-biome support for OpenLandLegacy

* Updated all the maps in OpenLandLegacy to use Multi Biome suppport.

* Add some changes with the QA result for OpenLandLegacy.